LU-2775 fid: allow FID-on-OST in fid_seq_is_mdt()
The old 2.1 LASSERT_SEQ_IS_MDT() macro used fid_seq_is_mdt() in
several places to verify that a FID was "sane" for where it was
being used.  However, fid_seq_is_mdt() only checked FID_SEQ_OST_MDT0
through FID_SEQ_OST_MDT9 values, which immediately broke 2.1 clients
when FID-on-OST was enabled on 2.4 MDTs.

The use of FID-on-OST will be limited to only DNE MDTs for 2.4, but
clients should not LASSERT() on a value that it got from the network.
The use of LASSERT_SEQ_IS_MDT() is removed from clients, since this
is "validating" data from the network.  Even with this removal, the
fid_seq_is_mdt() macro is updated to allow FID_SEQ_NORMAL sequences,
as in 2.4, in case it is needed elsewhere.

Reviewed-on: http://review.whamcloud.com/5304
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>

+++ lustre/obdfilter/filter.c
@@ -4517,11 +4517,11 @@ static int filter_set_mds_conn(struct obd_export *exp, void *val)
 
         /* setup llog imports */
         if (val != NULL)
-                group = (int)(*(__u32 *)val);
+                group = *(__u32 *)val;
         else
                 group = 0; /* default value */
 
-        LASSERT_SEQ_IS_MDT(group);
+        LASSERT_SEQ_IS_MDT((__u64)group);
         rc = filter_setup_llog_group(exp, obd, group);
         if (rc)
                 goto out;
